requirements.txt
aiogram==3.14.0
betterlogging==1.0.0
orjson==3.10.11
python-dotenv==1.0.1
beautifulsoup==4.12.3
certifi==2024.8.30
greenlet==1.5.0
notion-client==2.2.1
SQLAlchemy==2.0.36